Export createComment function
