--- js/post-creator.js
@@ -1,1 +1,0 @@
-// import { isGreaterThanMaxStringLength } from './util'
